+++ src/spring/azext_spring/_client_factory.py
@@ -32,14 +32,13 @@ from .vendored_sdks.appplatform.v2021_09_01_preview import (
 from .vendored_sdks.appplatform.v2022_09_01_preview import (
     AppPlatformManagementClient as AppPlatformManagementClient_20220901preview
 )
+from .vendored_sdks.appplatform.v2022_11_01_preview import (
+    AppPlatformManagementClient as AppPlatformManagementClient_20221101preview
+)
 
 
-def cf_spring_20220901preview(cli_ctx, *_):
-    return get_mgmt_service_client(cli_ctx, AppPlatformManagementClient_20220901preview)
-
-
-def cf_spring_20220901preview(cli_ctx, *_):
-    return get_mgmt_service_client(cli_ctx, AppPlatformManagementClient_20220901preview)
+def cf_spring_20221101preview(cli_ctx, *_):
+    return get_mgmt_service_client(cli_ctx, AppPlatformManagementClient_20221101preview)
 
 
 def cf_spring_20220901preview(cli_ctx, *_):