+++ src/spring/azext_spring/spring_instance.py
@@ -15,9 +15,11 @@ from ._tanzu_component import (create_application_configuration_service,
                                create_gateway,
                                create_api_portal)
 
-
 from ._validators import (_parse_sku_name, validate_instance_not_existed)
+from azure.cli.core.commands import LongRunningOperation
 from knack.log import get_logger
+from ._marketplace import _spring_list_marketplace_plan
+from ._constant import (MARKETPLACE_OFFER_ID, MARKETPLACE_PUBLISHER_ID)
 
 logger = get_logger(__name__)
 
@@ -62,6 +64,7 @@ class DefaultSpringCloud:
                        sku=None,
                        tags=None,
                        ingress_read_timeout=None,
+                       marketplace_plan_id=None,
                        **_):
         properties = models.ClusterResourceProperties(
             zone_redundant=zone_redundant
@@ -73,6 +76,13 @@ class DefaultSpringCloud:
             )
         else:
             properties.vnet_addons = None
+
+        if marketplace_plan_id:
+            properties.marketplace_resource = models.MarketplaceResource(
+                plan=marketplace_plan_id,
+                product=MARKETPLACE_OFFER_ID,
+                publisher=MARKETPLACE_PUBLISHER_ID
+            )
 
         if service_runtime_subnet or app_subnet or reserved_cidr_range:
             properties.network_profile = models.NetworkProfile(
@@ -95,8 +105,7 @@ class DefaultSpringCloud:
         poller = self.client.services.begin_create_or_update(
             self.resource_group, self.name, resource)
         logger.warning(" - Creating Service ..")
-        wait_till_end(self.cmd, poller)
-        return poller
+        return LongRunningOperation(self.cmd.cli_ctx)(poller)
 
 
 class EnterpriseSpringCloud(DefaultSpringCloud):
@@ -154,6 +163,7 @@ def spring_create(cmd, client, resource_group, name,
                   api_portal_instance_count=None,
                   enable_log_stream_public_endpoint=None,
                   ingress_read_timeout=None,
+                  marketplace_plan_id=None,
                   no_wait=False):
     """
     Because Standard/Basic tier vs. Enterprise tier creation are very different. Here routes the command to different
@@ -184,6 +194,7 @@ def spring_create(cmd, client, resource_group, name,
         'enable_api_portal': enable_api_portal,
         'api_portal_instance_count': api_portal_instance_count,
         'enable_log_stream_public_endpoint': enable_log_stream_public_endpoint,
+        'marketplace_plan_id': marketplace_plan_id,
         'no_wait': no_wait
     }
 
@@ -199,3 +210,7 @@ def _enable_app_insights(cmd, client, resource_group, name, location, app_insigh
     return create_default_buildpack_binding_for_application_insights(cmd, client, resource_group, name,
                                                                      location, app_insights_key, app_insights,
                                                                      sampling_rate)
+
+
+def spring_list_marketplace_plan(cmd, client):
+    return _spring_list_marketplace_plan(cmd, client)